+++ supportedBackends/sfs/src/test/scala/cromwell/backend/impl/sfs/config/MemoryDeclarationValidationSpec.scala
@@ -7,22 +7,22 @@ import org.scalatest.prop.TableDrivenPropertyChecks
 import org.scalatest.{FlatSpec, Matchers}
 import wdl4s.parser.MemoryUnit
 import wom.format.MemorySize
-import wom.values.{WomFloat, WomInteger}
+import wom.values.{WomFloat, WomLong}
 
 class MemoryDeclarationValidationSpec extends FlatSpec with Matchers with TableDrivenPropertyChecks {
   behavior of "MemoryDeclarationValidation"
 
   val validDeclaredAmounts = Table(
     ("declaration", "runtimeAmount", "expectedDefaultAmount", "expectedExtracted"),
-    ("Int memory", Option(2), None, Option(WomInteger(2 * 1000 * 1000 * 1000))),
-    ("Int memory_gb", Option(2), None, Option(WomInteger(2))),
+    ("Int memory", Option(2), None, Option(WomLong(2L * 1024L * 1024L * 1024L))),
+    ("Int memory_gb", Option(2), None, Option(WomLong(2))),
     ("Int memory_gb = 3", None, Option(3), None),
-    ("Int memory_gb = 3", Option(2), Option(3), Option(WomInteger(2))),
+    ("Int memory_gb = 3", Option(2), Option(3), Option(WomLong(2))),
     ("Int? memory_gb", None, None, None),
-    ("Int? memory_gb", Option(2), None, Option(WomInteger(2))),
+    ("Int? memory_gb", Option(2), None, Option(WomLong(2))),
     ("Int? memory_gb = 3", None, Option(3), None),
-    ("Int? memory_gb = 3", Option(2), Option(3), Option(WomInteger(2))),
-    ("Float memory", Option(2), None, Option(WomFloat(2 * 1000 * 1000 * 1000))),
+    ("Int? memory_gb = 3", Option(2), Option(3), Option(WomLong(2))),
+    ("Float memory", Option(2), None, Option(WomFloat(2F * 1024F * 1024F * 1024F))),
     ("Float memory_gb", Option(2), None, Option(WomFloat(2))),
     ("Float memory_gb = 3.0", None, Option(3), None),
     ("Float memory_gb = 3.0", Option(2), Option(3), Option(WomFloat(2))),
@@ -34,15 +34,15 @@ class MemoryDeclarationValidationSpec extends FlatSpec with Matchers with TableD
     ("Float? memory_gb = 3.0", Option(2), Option(3), Option(WomFloat(2))),
     ("Float? memory_gb = 3", None, Option(3), None),
     ("Float? memory_gb = 3", Option(2), Option(3), Option(WomFloat(2))),
-    ("Int memoryMin", Option(2), None, Option(WomInteger(2 * 1000 * 1000 * 1000))),
-    ("Int memoryMin_gb", Option(2), None, Option(WomInteger(2))),
+    ("Int memoryMin", Option(2), None, Option(WomLong(2L * 1024L * 1024L * 1024L))),
+    ("Int memoryMin_gb", Option(2), None, Option(WomLong(2))),
     ("Int memoryMin_gb = 3", None, Option(3), None),
-    ("Int memoryMin_gb = 3", Option(2), Option(3), Option(WomInteger(2))),
+    ("Int memoryMin_gb = 3", Option(2), Option(3), Option(WomLong(2))),
     ("Int? memoryMin_gb", None, None, None),
-    ("Int? memoryMin_gb", Option(2), None, Option(WomInteger(2))),
+    ("Int? memoryMin_gb", Option(2), None, Option(WomLong(2))),
     ("Int? memoryMin_gb = 3", None, Option(3), None),
-    ("Int? memoryMin_gb = 3", Option(2), Option(3), Option(WomInteger(2))),
-    ("Float memoryMin", Option(2), None, Option(WomFloat(2 * 1000 * 1000 * 1000))),
+    ("Int? memoryMin_gb = 3", Option(2), Option(3), Option(WomLong(2))),
+    ("Float memoryMin", Option(2), None, Option(WomFloat(2F * 1024F * 1024F * 1024F))),
     ("Float memoryMin_gb", Option(2), None, Option(WomFloat(2))),
     ("Float memoryMin_gb = 3.0", None, Option(3), None),
     ("Float memoryMin_gb = 3.0", Option(2), Option(3), Option(WomFloat(2))),
@@ -50,15 +50,15 @@ class MemoryDeclarationValidationSpec extends FlatSpec with Matchers with TableD
     ("Float? memoryMin_gb", Option(2), None, Option(WomFloat(2))),
     ("Float? memoryMin_gb = 3.0", None, Option(3), None),
     ("Float? memoryMin_gb = 3.0", Option(2), Option(3), Option(WomFloat(2))),
-    ("Int memoryMax", Option(2), None, Option(WomInteger(2 * 1000 * 1000 * 1000))),
-    ("Int memoryMax_gb", Option(2), None, Option(WomInteger(2))),
+    ("Int memoryMax", Option(2), None, Option(WomLong(2L * 1024L * 1024L * 1024L))),
+    ("Int memoryMax_gb", Option(2), None, Option(WomLong(2))),
     ("Int memoryMax_gb = 3", None, Option(3), None),
-    ("Int memoryMax_gb = 3", Option(2), Option(3), Option(WomInteger(2))),
+    ("Int memoryMax_gb = 3", Option(2), Option(3), Option(WomLong(2))),
     ("Int? memoryMax_gb", None, None, None),
-    ("Int? memoryMax_gb", Option(2), None, Option(WomInteger(2))),
+    ("Int? memoryMax_gb", Option(2), None, Option(WomLong(2))),
     ("Int? memoryMax_gb = 3", None, Option(3), None),
-    ("Int? memoryMax_gb = 3", Option(2), Option(3), Option(WomInteger(2))),
-    ("Float memoryMax", Option(2), None, Option(WomFloat(2 * 1000 * 1000 * 1000))),
+    ("Int? memoryMax_gb = 3", Option(2), Option(3), Option(WomLong(2))),
+    ("Float memoryMax", Option(2), None, Option(WomFloat(2F * 1024F * 1024F * 1024F))),
     ("Float memoryMax_gb", Option(2), None, Option(WomFloat(2))),
     ("Float memoryMax_gb = 3.0", None, Option(3), None),
     ("Float memoryMax_gb = 3.0", Option(2), Option(3), Option(WomFloat(2))),
@@ -96,7 +96,7 @@ class MemoryDeclarationValidationSpec extends FlatSpec with Matchers with TableD
       val extracted = memoryDeclarationValidation.extractWdlValueOption(validatedRuntimeAttributes)
 
       val expectedDefault = expectedDefaultAmount
-        .map(amount => WomInteger(MemorySize(amount.toDouble, MemoryUnit.GB).bytes.toInt))
+        .map(amount => WomLong(MemorySize(amount.toDouble, MemoryUnit.GB).bytes.toLong))
 
       MemoryDeclarationValidation.isMemoryDeclaration(runtimeDeclaration.unqualifiedName,
         memoryKey, memoryPrefix) should be(true)