+++ wom/src/main/java/wdl4s/parser/MemoryUnit.java
@@ -2,14 +2,10 @@ package wdl4s.parser;
 
 public enum MemoryUnit {
     Bytes(1, "B"),
-    KiB(1L << 10, "KiB", "Ki"),
-    MiB(1L << 20, "MiB", "Mi"),
-    GiB(1L << 30, "GiB", "Gi"),
-    TiB(1L << 40, "TiB", "Ti"),
-    KB(1000, "KB", "K"),
-    MB(KB.bytes * 1000, "MB", "M"),
-    GB(MB.bytes * 1000, "GB", "G"),
-    TB(GB.bytes * 1000, "TB", "T");
+    KB(1L << 10, "KB", "K", "KiB", "Ki"),
+    MB(1L << 20, "MB", "M", "MiB", "Mi"),
+    GB(1L << 30, "GB", "G", "GiB", "Gi"),
+    TB(1L << 40, "TB", "T", "TiB", "Ti");
 
     public final double bytes;
     public final String[] suffixes;